confirm delete message functionality and added styling

+++ public/client.js
@@ -16,8 +16,12 @@ function getListItems(){
       console.log('in getListItems: back from server-->', response);
       $('#list').empty();
       for (var i = 0; i < response.length; i++) {
-        $('#list').append('<div class="item" data-id='+response[i].id+'><button class="completed-button">Completed!</button>'+response[i].item+'<span></span><button class="delete-button">Delete Item</button></div>');
-      }
+        if(response[i].complete===true){
+          $('#list').append('<div class="item complete" data-id='+response[i].id+'><span><img src="/checkmark.png"/></span>'+response[i].item+'<span></span><button class="delete-button">Delete Item</button></div>');
+        } else{
+          $('#list').append('<div class="item" data-id='+response[i].id+'><span class="button-span"><button class="completed-button">Completed!</button></span>'+response[i].item+'<span></span><button class="delete-button">Delete Item</button></div>');
+        }  // end else
+      }  // end for loop
     }  //end success
   });  //end ajax
 }  // end getListItems
@@ -35,7 +39,7 @@ function addListItem(){
     data: listItemToSend,
     success: function(response){
       console.log('in addListItem: back from server with-->', response);
-      $('form').reset();
+      $('form').trigger('reset');
       getListItems();
     }  //end success
   });  //end ajax
@@ -43,39 +47,53 @@ function addListItem(){
 
 function deleteItem(){
   console.log('delete item button clicked');
-  var id = $(this).parent().data('id');
-  console.log('id:',id);
-  var idToSend = {
-    listId: id,
-  };
-  console.log('idToSend:', idToSend);
-  $.ajax({
-    type: 'DELETE',
-    url:'/deleteItem/',
-    data: idToSend,
-    success: function(response){
-      console.log('response is-->',response);
-      console.log('delete-button for id clicked');
-      getListItems();
-    }  // end success
-  });  // end ajax
-}
+  if (confirm('Did you really get this done?  Are you sure you want to delete this?')===true){
+    var id = $(this).parent().data('id');
+    console.log('id:',id);
+    var idToSend = {
+      listId: id,
+    };
+    console.log('idToSend:', idToSend);
+    $.ajax({
+      type: 'DELETE',
+      url:'/deleteItem/',
+      data: idToSend,
+      success: function(response){
+        console.log('response is-->',response);
+        console.log('delete-button for id clicked');
+        getListItems();
+      }  // end success
+    });  // end ajax
+  } else{
+    console.log('delete item cancelled');
+  }  // end else
+}  // end deleteItem
 
 function completeItem(){
   console.log('completed button clicked');
-  var id = $(this).parent().data('id');
+  var id = $(this).closest('div').data('id');
   console.log('id:',id);
   var idToSend = {
     listId: id,
   };
+  $(this).closest('div').addClass('complete');
+  $(this).closest('span').replaceWith('<span><img src="/checkmark.png"/></span>');
   console.log('idToSend:', idToSend);
   $.ajax({
     type: 'POST',
     url:'/completeItem/',
     data: idToSend,
     success: function(response){
-      console.log('response is-->',response);
-      console.log('completed-button for id clicked');
+      console.log('response from server is-->',response);
     }  // end success
   });  // end ajax
 }  // end completeItem
+
+function confirmDelete(){
+  if (confirm('Did you really get this done?  Are you sure you want to delete this?')===true){
+    deleteItem();
+  } else{
+    console.log('delete item cancelled');
+  }
+
+}
